[fixbugs] fix bugs in CityManager Enum

+++ app/src/main/java/com/hfad/weatherforecast/model/current/Precipitation.java
@@ -11,5 +11,5 @@ public class Precipitation {
 	public String mValue;
 
 	@SerializedName("day_value")
-	public int mDayValue;
+	public double mDayValue;
 }
